requirements.txt
Django==3.2.8
django-cors-headers==3.10.0
djangorestframework==3.12.4
Pillow==8.4.0
